fix logic of getting the roles ids

+++ migrations/000005_insert_roles.down.sql
@@ -1,10 +1,5 @@
 ALTER TABLE users ALTER COLUMN role_id DROP NOT NULL;
 
-UPDATE users SET role_id=NULL where role_id IN (1,2,3);
+UPDATE users SET role_id=NULL where role_id IN (select role_id from users limit 3);
 
-DELETE FROM roles
-	WHERE role_id=1;
-DELETE FROM roles
-	WHERE role_id=2;
-DELETE FROM roles
-	WHERE role_id=3;
+DELETE FROM roles WHERE role_id IN (select role_id from users limit 3);
